ori
Module app

Ori app module.

Macros

include_style

Include a style file.

Structs

App

The main application state.

AppBuilder

A builder for an App.

DelegateCx

The context passed to the [Delegate] trait.

StyleFileLoader

A style loader for a single file.

StyleLoader

A style loader, which can load styles from multiple sources.

WindowRenderState

Information needed to render a window.

Enums

AppCommand

Commands that can be sent to the application.

AppRequest

Requests that an application can make to the platform.

Traits

AppDelegate

A delegate for handling events in an application.

IntoUiBuilder

Trait for converting a type into a UiBuilder.

LoadStyle

A trait for loading styles.

Type Aliases

UiBuilder

A builder for a user interface.