ori::app
Struct App

pub struct App<T> {
    pub proxy: CommandProxy,
    pub contexts: Contexts,
    /* private fields */
}

The main application state.

Fields
proxy: CommandProxy

The command proxy of the application.

contexts: Contexts

The contexts of the application.

Implementations

impl<T> App<T>

pub fn build() -> AppBuilder<T>

Create a new application builder.

pub fn close_requested(&mut self, data: &mut T, window_id: WindowId) -> bool

A window was requested to be closed.

Returns true if the window was closed, i.e. the event was not handled.

pub fn window_resized( &mut self, data: &mut T, window_id: WindowId, width: u32, height: u32, ) -> bool

A window was resized.

pub fn window_scaled( &mut self, data: &mut T, window_id: WindowId, scale: f32, ) -> bool

A window was scaled.

pub fn window_maximized( &mut self, data: &mut T, window_id: WindowId, maximized: bool, ) -> bool

The maximized state of a window changed.

pub fn pointer_moved( &mut self, data: &mut T, window_id: WindowId, pointer_id: PointerId, position: Point, ) -> bool

A pointer moved.

pub fn pointer_left( &mut self, data: &mut T, window_id: WindowId, pointer_id: PointerId, ) -> bool

A pointer left the window.

pub fn pointer_scrolled( &mut self, data: &mut T, window_id: WindowId, pointer_id: PointerId, delta: Vector, ) -> bool

A pointer scrolled.

pub fn pointer_button( &mut self, data: &mut T, window_id: WindowId, pointer_id: PointerId, button: PointerButton, pressed: bool, ) -> bool

A pointer button was pressed or released.

pub fn keyboard_key( &mut self, data: &mut T, window_id: WindowId, key: Key, code: Option<Code>, text: Option<String>, pressed: bool, ) -> bool

A keyboard key was pressed or released.

pub fn modifiers_changed(&mut self, modifiers: Modifiers)

The modifiers changed.

impl<T> App<T>

pub fn add_window( &mut self, data: &mut T, ui: Box<dyn FnMut(&mut T) -> Box<dyn AnyView<T>>>, window: Window, )

Add a window to the application.

pub fn remove_window(&mut self, window_id: WindowId)

Remove a window from the application.

pub fn get_window(&self, window_id: WindowId) -> Option<&Window>

Get a window by id.

pub fn get_window_mut(&mut self, window_id: WindowId) -> Option<&mut Window>

Get a mutable window by id.

pub fn add_context(&mut self, context: impl Any)

Add a context.

pub fn take_requests(&mut self) -> impl Iterator<Item = AppRequest<T>>

Take all pending requests.

pub fn handle_commands(&mut self, data: &mut T)

Handle all pending commands.

pub fn update_hovered(&mut self, window_id: WindowId) -> bool

Update the hovered state of a window.

pub fn init(&mut self, data: &mut T)

Initialize the application.

pub fn idle(&mut self, data: &mut T)

The application is idle.

pub fn rebuild(&mut self, data: &mut T)

Rebuild all windows.

pub fn event(&mut self, data: &mut T, event: &Event) -> bool

Handle an event for the entire application.

Returns true if the event was handled by a delegate.

pub fn window_event( &mut self, data: &mut T, window_id: WindowId, event: &Event, ) -> bool

Handle an event for a single window.

Returns true if the event was handled by a delegate.

pub fn draw_window( &mut self, data: &mut T, window_id: WindowId, ) -> Option<WindowRenderState>

Draw a single window, returning the scene if it needs to be rendered.