Struct ori::core::canvas::Color

#[repr(C)]
pub struct Color {
    pub r: f32,
    pub g: f32,
    pub b: f32,
    pub a: f32,
}

A color with red, green, blue and alpha components.

Fields
r: f32

The red component of the color.

g: f32

The green component of the color.

b: f32

The blue component of the color.

a: f32

The alpha component of the color.

Implementations

impl Color

pub const TRANSPARENT: Color = _

pub const BLACK: Color = _

pub const WHITE: Color = _

pub const RED: Color = _

pub const GREEN: Color = _

pub const BLUE: Color = _

pub const YELLOW: Color = _

pub const CYAN: Color = _

pub const MAGENTA: Color = _

impl Color

pub const fn rgba(r: f32, g: f32, b: f32, a: f32) -> Color

Create a new color with the given red, green, blue and alpha components.

pub const fn rgb(r: f32, g: f32, b: f32) -> Color

Create a new color with the given red, green and blue components.

pub const fn grayscale(g: f32) -> Color

Create a new color with the given gray component.

pub fn rgba8(r: u8, g: u8, b: u8, a: u8) -> Color

Create a new color with the given red, green, blue and alpha components.

pub fn r8(&self) -> u8

Get the red component as an 8 bit integer.

pub fn g8(&self) -> u8

Get the green component as an 8 bit integer.

pub fn b8(&self) -> u8

Get the blue component as an 8 bit integer.

pub fn a8(&self) -> u8

Get the alpha component as an 8 bit integer.

pub fn try_hex(hex: &str) -> Option<Color>

Try to parse a color from a hex string.

pub fn hex(hex: &str) -> Color

Parse a color from a hex string.

Panics

If the string is not a valid hex color.

pub fn to_hex(self) -> DisplayHex

Convert the color to hex.

pub fn hsla(h: f32, s: f32, l: f32, a: f32) -> Color

Returns a new color with the given hue, saturation, lightness and alpha components.

See https://en.wikipedia.org/wiki/HSL_and_HSV.

pub fn hsl(h: f32, s: f32, l: f32) -> Color

Returns a new color with the given hue, saturation, lightness and alpha components.

See https://en.wikipedia.org/wiki/HSL_and_HSV.

pub fn to_hsla(self) -> (f32, f32, f32, f32)

Convert the color to a hue, saturation, lightness and alpha tuple.

See https://en.wikipedia.org/wiki/HSL_and_HSV.

pub fn to_hsl(self) -> (f32, f32, f32)

Convert the color to a hue, saturation, lightness tuple.

See https://en.wikipedia.org/wiki/HSL_and_HSV.

pub fn hsva(h: f32, s: f32, v: f32, a: f32) -> Color

Returns a new color with the given hue, saturation, value and alpha components.

See https://en.wikipedia.org/wiki/HSL_and_HSV.

pub fn hsv(h: f32, s: f32, v: f32) -> Color

Returns a new color with the given hue, saturation and value.

See https://en.wikipedia.org/wiki/HSL_and_HSV.

pub fn to_hsva(self) -> (f32, f32, f32, f32)

Convert the color to a hue, saturation, value and alpha tuple.

See https://en.wikipedia.org/wiki/HSL_and_HSV.

pub fn to_hsv(self) -> (f32, f32, f32)

Convert the color to a hue, saturation, value tuple.

See https://en.wikipedia.org/wiki/HSL_and_HSV.

pub fn oklaba(l: f32, a: f32, b: f32, alpha: f32) -> Color

Convert a color from oklab to sRGB.

See https://bottosson.github.io/posts/oklab/.

pub fn oklab(l: f32, a: f32, b: f32) -> Color

Convert a color from oklab to sRGB.

See https://bottosson.github.io/posts/oklab/.

pub fn to_oklaba(self) -> (f32, f32, f32, f32)

Convert a color from sRGB to oklab.

See https://bottosson.github.io/posts/oklab/.

pub fn to_oklab(self) -> (f32, f32, f32)

Convert a color from sRGB to oklab.

See https://bottosson.github.io/posts/oklab/.

pub fn oklcha(l: f32, c: f32, h: f32, alpha: f32) -> Color

Convert a color from oklch to sRGB.

See https://bottosson.github.io/posts/oklab/.

pub fn oklch(l: f32, c: f32, h: f32) -> Color

Convert a color from oklch to sRGB.

See https://bottosson.github.io/posts/oklab/.

pub fn to_oklcha(self) -> (f32, f32, f32, f32)

Convert a color from sRGB to oklch.

See https://bottosson.github.io/posts/oklab/.

pub fn to_oklch(self) -> (f32, f32, f32)

Convert a color from sRGB to oklch.

See https://bottosson.github.io/posts/oklab/.

pub fn okhsla(h: f32, s: f32, l: f32, alpha: f32) -> Color

Convert a color from okhsl to sRGB.

pub fn okhsl(h: f32, s: f32, l: f32) -> Color

Convert a color from okhsl to sRGB.

pub fn to_okhsla(self) -> (f32, f32, f32, f32)

Convert a color from sRGB to okhsl.

pub fn to_okhsl(self) -> (f32, f32, f32)

Convert a color from sRGB to okhsl.

pub fn okhsva(h: f32, s: f32, v: f32, alpha: f32) -> Color

Convert a color from okhsv to sRGB.

pub fn okhsv(h: f32, s: f32, v: f32) -> Color

Convert a color from okhsv to sRGB.

pub fn to_okhsva(self) -> (f32, f32, f32, f32)

Convert a color from sRGB to okhsv.

pub fn to_okhsv(self) -> (f32, f32, f32)

Convert a color from sRGB to okhsv.

pub fn luminocity(self) -> f32

Get the luminocity.

pub fn mix(self, other: Color, t: f32) -> Color

Linearly interpolate between two colors.

This uses a fractor t between 0.0 and 1.0. Where 0.0 is self and 1.0 is other.

Note that this is a linear interpolation in the oklab color space. If rgb interpolation is required use mix_rgb.

pub fn mix_rgb(self, other: Color, t: f32) -> Color

Linearly interpolate between two colors.

This uses a fractor t between 0.0 and 1.0. Where 0.0 is self and 1.0 is other.

Note that this is a linear interpolation in the sRGB color space. If this isn’t necessary use mix, as it uses the oklab color space, which is more perceptually uniform.

pub fn saturate(self, amount: f32) -> Color

Saturates the color by given amount.

pub fn desaturate(self, amount: f32) -> Color

Desaturates the color by given amount.

pub fn lighten(self, amount: f32) -> Color

Brighten the color by the given amount.

pub fn darken(self, amount: f32) -> Color

Darken the color by the given amount.

pub fn fade(self, amount: f32) -> Color

Fade the color by the given amount.

pub fn is_translucent(self) -> bool

Returns true if the color is translucent.

pub fn is_transparent(self) -> bool

Returns true if the color is transparent.

pub fn to_srgb(self) -> [f32; 4]

Convert the color to linear sRGB.

See https://en.wikipedia.org/wiki/SRGB.

pub fn to_rgba8(self) -> [u8; 4]

Convert the color to linear sRGB.