ori::core::canvas
Struct Curve

pub struct Curve { /* private fields */ }

A bezier curve.

Implementations

impl Curve

pub fn new() -> Curve

Create a new curve.

pub fn rect(rect: Rect) -> Curve

Create a curve from a rectangle.

pub fn ellipse(oval: Rect) -> Curve

Create a curve from an oval.

pub fn circle(center: Point, radius: f32) -> Curve

Create a curve from a cicrle.

pub fn len(&self) -> usize

Get the number of verbs in the curve.

pub fn is_empty(&self) -> bool

Check if the curve is empty.

pub fn is_valid(&self) -> bool

Check if the curve is valid.

pub fn is_closed(&self) -> bool

Check if the curve is closed.

pub fn bounds(&self) -> Rect

Get the bounds of the curve.

pub fn clear(&mut self)

Clear the curve, retaining the allocated memory for reuse.

pub fn points(&self) -> &[Point]

Get the points in the curve.

pub fn verbs(&self) -> &[CurveVerb]

Get the verbs in the curve.

pub fn last_point(&self) -> Option<Point>

Get the last point in the curve.

pub fn move_to(&mut self, point: Point)

Move to a point.

pub fn line_to(&mut self, point: Point)

Draw a line to a point.

pub fn quad_to(&mut self, control: Point, point: Point)

Draw a quadratic bezier curve to a point, with a control point control.

pub fn cubic_to(&mut self, a: Point, b: Point, point: Point)

Draw a cubic bezier curve to a point, with control points a and b.

pub fn conic_to(&mut self, control: Point, point: Point, weight: f32)

Draw a conic curve to a point, with a control point control and a weight.

Conic curves are approximated by quadratic bezier curves.

pub fn close(&mut self)

Close the contour.

pub fn transform(&mut self, transform: Affine)

Transform the curve by the given affine transform.

pub fn iter(&self) -> CurveIter<'_>

Get an iterator over the curve segments.

pub fn contains(&self, point: Point, rule: FillRule) -> bool

Check if the curve contains a point using the given rule.

pub fn stroke_curve(&mut self, curve: &Curve, stroke: Stroke)

Stroke the curve with the given stroke.

pub fn push_rect(&mut self, rect: Rect)

Push a rectangle to the curve.

pub fn push_oval(&mut self, oval: Rect)

Push an oval to the curve.

pub fn push_rect_with_radius(&mut self, rect: Rect, radius: BorderRadius)

Push a rectangle with rounded corners to the curve.

pub fn push_rect_with_borders( &mut self, rect: Rect, radius: BorderRadius, width: BorderWidth, )

Push the border of a rectangle with rounded corners to the curve.