ori::core
Module command

A channel for sending commands to the user interface.

Structs

Command

A command containing arbitrary data.

CommandProxy

A clonable channel for sending Commands.

CommandReceiver

A receiver for Commands.

CommandWaker

A waker for the event loop, triggered when a command is sent.