ori::core
Module event

Events and event handling.

Structs

Ime

Input Method Editor (IME) state.

KeyPressed

An event fired when a key is pressed.

KeyReleased

An event fired when a key is released.

Modifiers

Modifiers for keyboard and pointer events.

PointerId

A unique pointer id.

PointerLeft

A pointer left the window.

PointerMoved

A pointer was moved.

PointerPressed

A pointer button was pressed.

PointerReleased

A pointer button was released.

PointerScrolled

A pointer wheel was scrolled.

RequestFocus

A request to focus a view.

RequestFocusNext

A request to focus the next view in the focus chain.

RequestFocusPrev

A request to focus the previous view in the focus chain.

WindowCloseRequested

Event emitted when a window wants to close.

WindowMaximized

Event emitted when a window is maximized.

WindowResized

Event emitted when a window is resized.

WindowScaled

Event emitted when a window is scaled.

Enums

Capitalize

Input Method Editor (IME) capitalization.

Code

A keyboard key-code.

Event

An event that can be sent to a view.

FocusTarget

A target for focus.

Key

A keyboard key.

PointerButton

A pointer button.

Traits

IsKey

A trait for checking if something is a certain key.