ori::core::layout
Struct Rect

pub struct Rect {
    pub min: Point,
    pub max: Point,
}

A rectangle defined by its minimum and maximum points.

Fields
min: Point

The minimum point of the rectangle.

max: Point

The maximum point of the rectangle.

Implementations

impl Rect

pub const ZERO: Rect = _

A rectangle with zero area.

pub const fn new(min: Point, max: Point) -> Rect

Create a new rectangle with the given minimum and maximum points.

pub fn min_size(min: Point, size: Size) -> Rect

Create a new rectangle with the given minimum point and size.

pub fn max_size(max: Point, size: Size) -> Rect

Create a new rectangle with the given maximum point and size.

pub fn center_size(center: Point, size: Size) -> Rect

Create a new rectangle with the given center point and size.

pub fn round(self) -> Rect

Round the rectangle’s minimum point down and its maximum point up.

pub fn clamp(self, other: impl Into<Rect>) -> Rect

Clamp the rectangle to the given rectangle.

pub fn offset(self) -> Vector

Get the offset of the rectangle.

pub fn size(self) -> Size

Get the size of the rectangle.

pub fn width(self) -> f32

Get the width of the rectangle.

pub fn height(self) -> f32

Get the height of the rectangle.

pub fn area(self) -> f32

Get the area of the rectangle.

pub fn top_left(self) -> Point

Get the top left point of the rectangle.

pub fn top_center(self) -> Point

Get the top center point of the rectangle.

pub fn top_right(self) -> Point

Get the top right point of the rectangle.

pub fn center_left(self) -> Point

Get the left center point of the rectangle.

pub fn center(self) -> Point

Get the center point of the rectangle.

pub fn center_right(self) -> Point

Get the right center point of the rectangle.

pub fn bottom_left(self) -> Point

Get the bottom left point of the rectangle.

pub fn bottom_center(self) -> Point

Get the bottom center point of the rectangle.

pub fn bottom_right(self) -> Point

Get the bottom right point of the rectangle.

pub fn top(self) -> f32

Get the top edge of the rectangle.

pub fn left(self) -> f32

Get the left edge of the rectangle.

pub fn right(self) -> f32

Get the right edge of the rectangle.

pub fn bottom(self) -> f32

Get the bottom edge of the rectangle.

pub fn is_finite(self) -> bool

Check if the rectangle is finite.

pub fn is_infinite(self) -> bool

Check if the rectangle is infinite.

pub fn shrink(self, padding: f32) -> Rect

Shrink the rectangle by the given amount.

pub fn expand(self, padding: f32) -> Rect

Expand the rectangle by the given amount.

pub fn contains(self, point: Point) -> bool

Compute whether the rectangle contains the given point.

pub fn contain(self, point: Point) -> Point

Compute the closest point in the rectangle to the given point.

pub fn include(self, point: Point) -> Rect

Expand the rectangle to contain the given point.

pub fn try_intersection(self, other: Rect) -> Option<Rect>

Compute the intersection of the rectangle with the given rectangle.

pub fn intersection(self, other: Rect) -> Rect

Compute the intersection of the rectangle with the given rectangle.

If the rectangles do not intersect Rect::ZERO is returned.

pub fn intersects(self, other: Rect) -> bool

Check if the rectangle intersects the given rectangle.

pub fn union(self, other: Rect) -> Rect

Compute the union of the rectangle with the given rectangle.

pub fn transform(self, transform: Affine) -> Rect

Transform the rectangle by the given affine transform.