Module ori::core::prelude

Convenient imports for Ori.

Macros

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

hstack

Create a horizontal Stack.

hwrap

Create a horizontal Wrap.

include_font

Load a font from a file or directory.

info

Constructs an event at the info level.

text

Create a formatted Text.

trace

Constructs an event at the trace level.

vstack

Create a vertical Stack.

vwrap

Create a vertical Wrap.

warn

Constructs an event at the warn level.

zstack

Create a new ZStack view.

Structs

Affine

An affine transformation in 2 dimensional space.

Aligned

A view that aligns its content.

Alignment

Alignment of content inside a container.

Animate

A view that animates.

Background

A background.

BorderRadius

Radi of the corners on a Quad.

BorderWidth

The border width of a Quad.

BoxShadow

A box shadow.

BuildCx

A context for building the view tree.

BuildHandler

A view that hooks into the build cycle.

Button

A button.

ButtonStyle

The style of a button.

Canvas

A canvas used for drawing a Scene.

Checkbox

A checkbox.

CheckboxStyle

The style of a Checkbox.

Clickable

A click handler.

CloseRequested

Event emitted when a window wants to close.

Collapsing

A collapsing view.

CollapsingStyle

The style of a collapsing view.

Color

A color with red, green, blue and alpha components.

CommandProxy

A clonable channel for sending Commands.

Constrain

A view that constrains its content to a given space.

Container

A container view.

ContainerStyle

The style of a Container.

Curve

A curve.

Depth

A view with a depth value.

DrawCx

A context for drawing the view tree.

DrawHandler

A view that hooks into the draw cycle.

Dropdown

A dropdown view.

EventCx

A context for handling events.

EventHandler

A view that handles events.

Flex

A flexible view.

Focus

A view that focuses on a part of the data.

FontWeight

A font weight.

Fonts

A context for loading and rasterizing fonts.

Fragment

A item to be rendered.

Image

An clonable image.

ImageData

Image data.

ImageId

A unique identifier for an Image.

KeyPressed

An event fired when a key is pressed.

LayoutBuilder

A view that builds its content based on the layout constraints.

LayoutCx

A context for laying out the view tree.

Matrix

A 2x2 matrix.

Memo

A view that only builds the inner view when certain data changes.

Mesh

A mesh containing vertices, indices and an optional image.

Modifiers

Modifiers for keyboard and pointer events.

Opaque

A view that bridges the gap between impl View and impl View<T>.

Pad

A view that adds padding to its content.

Padding

A padding of a rectangle.

Painter

A view that draws something.

Palette

A color palette.

Pod

A view that has separate ViewState from its content.

PodSeq

Contents of a view, in a sequence.

Point

A point in 2D space.

Pointer

The state of a pointer.

PointerId

A unique pointer id.

PointerMoved

A pointer was moved.

PointerPressed

A pointer button was pressed.

PointerReleased

A pointer button was released.

PointerScrolled

A pointer wheel was scrolled.

RebuildCx

A context for rebuilding the view tree.

RebuildHandler

A view that handles rebuilds.

Rect

A rectangle defined by its minimum and maximum points.

Scroll

A scrollable view.

SeqState

The state of a PodSeq.

Size

A 2 dimensional size.

Slider

A slider.

SliderStyle

The style of a slider.

Space

Space available to lay out a view.

Stack

A view that stacks it’s content in a line.

State

The state of a Pod.

Styles

A map of style values.

Text

A view that displays text.

TextBuffer

A convenient wrapper around a [cosmic_text::Buffer].

TextInput

A text input.

TextInputStyle

The style of a text input.

TextStyle

The style of a text.

Tooltip

A view that displays some text when the content is hovered.

TooltipStyle

The style of a tooltip.

Transform

A view that transforms its content.

Transition

A transition.

Trigger

A view that creates a trigger around the content.

Vector

A 2D vector.

Vertex

A vertex in a Mesh.

ViewState

State associated with a View.

Window

A window.

WindowId

A unique identifier for a window.

WithState

A view that stores some additional data.

Wrap

A view that lays out it’s content in a line wrapping if it doesn’t fit.

ZStack

A view that overlays its content on top of each other.

Enums

Align

The alignment of items along the cross axis.

Axis

An axis is a direction in which a layout is applied.

Code

A keyboard key.

Cursor

A mouse cursor icon.

Easing

A transition easing curve.

Event

An event that can be sent to a view.

FontFamily

A font family, by default FontFamily::SansSerif.

FontSource

A source for a font.

FontStretch

A font stretch.

FontStyle

A font style.

Justify

The justify content of a stack container.

PointerButton

A pointer button.

Primitive

A primitive to be rendered.

TextAlign

Alignment of a section of text.

TextWrap

Wrapping of a section of text.

WindowSizing

The sizing of a window.

Constants

FILLDeprecation planned

Infinity (∞). Use f32::INFINITY instead.

Traits

AnyView

A view that supports dynamic dispatch.

Clipboard

A clipboard.

Rebuild

A convenience trait for rebuilding a View.

Style

A trait for styling a value.

View

A single UI component.

ViewSeq

A sequence of views.

Functions

align

Create a new Aligned view.

animate

Create a new Animate.

any

Create a new BoxedView.

background

Create a new Container with background.

bottom

Create a new Aligned view that aligns its content to the bottom.

bottom_left

Create a new Aligned view that aligns its content to the bottom left.

bottom_right

Create a new Aligned view that aligns its content to the bottom right.

button

Create a new Button.

center

Create a new Aligned view that aligns its content to the center.

checkbox

Create a new Checkbox.

clickable

Create a new Clickable.

collapsing

Create a new Collapsing.

constrain

Create a new Constrained view, constraining its content to a space.

container

Create a new Container.

depth

Create a new Depth view.

dropdown

Create a new Dropdown view.

ease

Create an ease transition with the given duration.

expand

Create a new expanded Flex view.

flex

Create a new Flex view.

focus

Create a new Focus.

gradient

Create a new gradient image.

height

Create a new Constrained view, constraining its content to a height.

hex

Create a new color, with the given hex string.

hscroll

Create a new horizontal Scroll.

hsl

Create a new color, with the given hue, saturation and lightness components.

hsla

Create a new color, with the given hue, saturation, lightness and alpha components.

hstack

Create a horizontal Stack.

hstack_any

Create a horizontal Stack, with dynamic content.

hstack_vec

Create a horizontal Stack, with vector content.

hwrap

Create a horizontal Wrap.

hwrap_any

Create a horizontal Wrap, with dynamic content.

hwrap_vec

Create a horizontal Wrap, with a vector of content.

layout_builder

Create a new LayoutBuilder view.

left

Create a new Aligned view that aligns its content to the left.

linear

Create a linear transition with the given duration.

max_height

Create a new Constrained view, constraining its content to a maximum height.

max_size

Create a new Constrained view, constraining its content to a maximum size.

max_width

Create a new Constrained view, constraining its content to a maximum width.

memo

Create a new Memo.

min_height

Create a new Constrained view, constraining its content to a minimum height.

min_size

Create a new Constrained view, constraining its content to a minimum size.

min_width

Create a new Constrained view, constraining its content to a minimum width.

oklab

Create a new color, with the given lightness, a and b components.

oklaba

Create a new color, with the given lightness, a, b and alpha components.

on_build

Create a new BuildHandler.

on_click

Create a new Clickable, with an on_click callback.

on_draw

Create a new DrawHandler.

on_event

Create a new EventHandler, with an after callback.

on_event_before

Create a new EventHandler, with a before callback.

on_press

Create a new Clickable, with an on_press callback.

on_rebuild

Create a new RebuildHandler.

on_release

Create a new Clickable, with an on_release callback.

opaque

A view that bridges the gap between impl View and impl View<T>.

pad

Create a new Pad view.

pad_bottom

Create a new Pad view adding padding to the bottom.

pad_left

Create a new Pad view adding padding to the left.

pad_right

Create a new Pad view adding padding to the right.

pad_top

Create a new Pad view adding padding to the top.

painter

Create a new Painter view.

palette

Get the palette of the style.

pod

Create a new Pod view.

rgb

Create a new color, with the given red, green and blue components.

rgba

Create a new color, with the given red, green, blue and alpha components.

right

Create a new Aligned view that aligns its content to the right.

rotate

Create a new Transform view that rotates its content.

rotate_degrees

Create a new Transform view that rotates its content in degrees.

scale

Create a new Transform view that scales its content.

shadow

Create a new Container with shadow.

size

Create a new Constrained view, cosntraining its content to a size.

slider

Create a new Slider.

style

Get a value from the current style.

style_or

Get a value from the current style or a default value.

styled

Run a closure with the given style.

text

Create a new Text.

text_input

Create a new TextInput.

tooltip

Create a new Tooltip view.

top

Create a new Aligned view that aligns its content to the top.

top_left

Create a new Aligned view that aligns its content to the top left.

top_right

Create a new Aligned view that aligns its content to the top right.

transform

Create a new Transform view.

transition

Animate a transition.

transition_active

Animate a view when active changes.

transition_focused

Animate a view when focused changes.

transition_hot

Animate a view when hot changes.

translate

Create a new Transform view that translates its content.

trigger

Create a new Trigger view.

vscroll

Create a new vertical Scroll.

vstack

Create a vertical Stack.

vstack_any

Create a vertical Stack, with dynamic content.

vstack_vec

Create a vertical Stack, with vector content.

vwrap

Create a vertical Wrap.

vwrap_any

Create a vertical Wrap, with dynamic content.

vwrap_vec

Create a vertical Wrap, with a vector of content.

width

Create a new Constrained view, constraining its content to a width.

with_state

Create a new WithState.

with_state_default

Create a new WithState using S::default().

without_data

Create a new view unwrapping some data from the state.

without_state

Create a new view unwrapping some state from the data.

zstack

Create a new ZStack view.

Type Aliases

BoxedView

A boxed dynamic view.

Lens

A lens used by Focus.

Derive Macros

Rebuild

Derive the Rebuild trait.