Enum ori::core::prelude::Event

#[non_exhaustive]
pub enum Event {

    WindowResized(WindowResized),
    CloseRequested(CloseRequested),
    WindowScaled(WindowScaled),
    PointerMoved(PointerMoved),
    PointerLeft(PointerLeft),
    PointerPressed(PointerPressed),
    PointerReleased(PointerReleased),
    PointerScrolled(PointerScrolled),
    KeyPressed(KeyPressed),
    KeyReleased(KeyReleased),
    Animate(f32),
    Command(Command),
    Update,

}

An event that can be sent to a view.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

WindowResized(WindowResized)

The window was resized.

CloseRequested(CloseRequested)

The window requested to be close.

WindowScaled(WindowScaled)

The window was scaled.

PointerMoved(PointerMoved)

A pointer moved.

PointerLeft(PointerLeft)

A pointer left the window.

PointerPressed(PointerPressed)

A pointer button was pressed.

PointerReleased(PointerReleased)

A pointer button was released.

PointerScrolled(PointerScrolled)

A pointer was scrolled.

KeyPressed(KeyPressed)

A keyboard key was pressed.

KeyReleased(KeyReleased)

A keyboard key was released.

Animate(f32)

An animation frame has passed.

Command(Command)

A command was sent.

Update

View state needs to be updated.

Implementations

impl Event

pub fn is_cmd<T>(&self) -> bool
where T: Any,

Check if the event is a command of a specific type.

pub fn cmd<T>(&self) -> Option<&T>
where T: Any,

Try to get the command as a specific type.

Returns None if the event is not a command or if the command is not of the specified type.