Struct ori::core::prelude::Vector

#[repr(C)]
pub struct Vector {
    pub x: f32,
    pub y: f32,
}

A 2D vector.

Fields
x: f32

The x coordinate.

y: f32

The y coordinate.

Implementations

impl Vector

pub const ZERO: Vector = _

The zero vector.

pub const ONE: Vector = _

The one vector.

pub const X: Vector = _

The unit x vector.

pub const Y: Vector = _

The unit y vector.

pub const NEG_X: Vector = _

The negative unit x vector.

pub const NEG_Y: Vector = _

The negative unit y vector.

pub const fn new(x: f32, y: f32) -> Vector

Create a new vector.

pub const fn all(value: f32) -> Vector

Create a new vector with the same x and y.

pub fn from_angle(angle: f32) -> Vector

Create a new vector from an angle.

let x = std::f32::consts::FRAC_PI_3;
let v = Vector::from_angle(x);

assert_eq!(v.x, x.cos());
assert_eq!(v.y, x.sin());

pub fn min(self, other: Vector) -> Vector

Get the min of self and other by element.

pub fn max(self, other: Vector) -> Vector

Get the max of self and other by element.

pub fn clamp(self, min: Vector, max: Vector) -> Vector

Clamp self to the range [min, max] by element.

pub fn floor(self) -> Vector

Floor the vector by element.

pub fn ceil(self) -> Vector

Ceil the vector by element.

pub fn round(self) -> Vector

Round the vector by element.

pub fn signum(self) -> Vector

Get the absolute value of the vector.

pub fn length_squared(self) -> f32

Get the length of the vector squared.

pub fn length(self) -> f32

Get the length of the vector.

pub fn normalize(self) -> Vector

Normalize the vector.

If the length of the vector is zero, the zero vector is returned.

pub fn dot(self, other: Vector) -> f32

Get the dot product of self and other.

pub fn cross(self, other: Vector) -> f32

Get the length of the cross product of self and other.

pub fn hat(self) -> Vector

Hat the vector.

pub fn angle_between(self, other: Vector) -> f32

Get the angle between self and other.

pub fn angle(self) -> f32

Get the angle of the vector.

pub const fn to_point(self) -> Point

Convert the vector to a vector.

pub const fn to_size(self) -> Size

Convert the vector to a size.