ori::core::view
Struct Pod

pub struct Pod<V> { /* private fields */ }

A view that has separate ViewState from its content.

When calling for example View::event, an EventCx is passed to the function. This EventCx contains a mutable reference to a ViewState that is used to keep track of state like whether the view is hovered or active. If a pod is not used when implementing a view, the View and the content share the same ViewState. This is almost always an issue when the View wants to have a diffrent transform or size than the content. See for example the Pad view.

Examples

use ori::prelude::*;

struct ContainerView<V> {
    // We wrap the content in a Pod here
    content: Pod<V>,
}

impl<V: View<T>, T> View<T> for ContainerView<V> {
    // We use the Pod's state here
    type State = State<T, V>;

    fn build(&mut self, cx: &mut BuildCx, data: &mut T) -> Self::State {
        self.content.build(cx, data)
    }

    // ...
}

Implementations

impl<V> Pod<V>

pub const fn new(view: V) -> Pod<V>

Create a new pod view.

pub fn event_maybe<T>( &mut self, handled: bool, state: &mut State<T, V>, cx: &mut EventCx<'_, '_>, data: &mut T, event: &Event, ) -> bool
where V: View<T>,

Call the View::event method on the content, only if the event hasn’t been handled.