Struct ori::core::view::RebuildCx

pub struct RebuildCx<'a, 'b> { /* private fields */ }

A context for rebuilding the view tree.

Implementations

impl<'a, 'b> RebuildCx<'a, 'b>

pub fn child(&mut self) -> RebuildCx<'_, 'b>

Create a child context.

pub fn build_cx(&mut self) -> BuildCx<'_, 'b>

Get a build context.

pub fn layout_cx(&mut self) -> LayoutCx<'_, 'b>

Get a layout context.

impl RebuildCx<'_, '_>

pub fn size(&self) -> Size

Get the size of the view.

pub fn rect(&self) -> Rect

Get the rect of the view in local space.

impl RebuildCx<'_, '_>

pub fn request_animation_frame(&mut self)

Request an animation frame.

pub fn window(&mut self) -> &mut Window

Get the window.

pub fn rasterize_text(&mut self, buffer: &TextBuffer) -> Mesh

Create a mesh for the given text buffer.

pub fn rasterize_text_raw(&mut self, buffer: &Buffer) -> Mesh

Create a mesh for the given raw cosmic text buffer.

pub fn id(&self) -> ViewId

Get the id of the view.

pub fn is_hot(&self) -> bool

Get whether the view is hot.

pub fn set_hot(&mut self, hot: bool) -> bool

Set whether the view is hot.

Returns true if the hot state changed.

pub fn is_focused(&self) -> bool

Get whether the view is focused.

pub fn set_focused(&mut self, focused: bool) -> bool

Set whether the view is focused.

Returns true if the focused state changed.

pub fn is_active(&self) -> bool

Get whether the view is active.

pub fn set_active(&mut self, active: bool) -> bool

Set whether the view is active.

Returns true if the active state changed.

pub fn has_hot(&self) -> bool

Get whether a child view is hot.

pub fn has_focused(&self) -> bool

Get whether a child view is focused.

pub fn has_active(&self) -> bool

Get whether a child view is active.

pub fn flex(&self) -> f32

Get the flex of the view.

pub fn set_flex(&mut self, flex: f32)

Set the flex of the view.

pub fn is_tight(&self) -> bool

Get whether the view is tight.

pub fn set_tight(&mut self, tight: bool)

Set whether the view is tight.

pub fn request_rebuild(&mut self)

Request a rebuild of the view tree.

pub fn request_layout(&mut self)

Request a layout of the view tree.

pub fn request_draw(&mut self)

Request a draw of the view tree.