Struct ori::core::view::ViewState

pub struct ViewState { /* private fields */ }

State associated with a View.

Implementations

impl ViewState

pub fn id(&self) -> ViewId

Get the id of the view.

pub fn is_hot(&self) -> bool

Get whether the view is hot.

pub fn set_hot(&mut self, hot: bool)

Set whether the view is hot.

pub fn is_focused(&self) -> bool

Get whether the view is focused.

pub fn set_focused(&mut self, focused: bool)

Set whether the view is focused.

pub fn is_active(&self) -> bool

Get whether the view is active.

pub fn set_active(&mut self, active: bool)

Set whether the view is active.

pub fn has_hot(&self) -> bool

Get whether the view has a hot child.

pub fn has_focused(&self) -> bool

Get whether the view has a focused child.

pub fn has_active(&self) -> bool

Get whether the view has an active child.

pub fn is_focusable(&self) -> bool

Get whether the view is focusable.

pub fn set_focusable(&mut self, focusable: bool)

Set whether the view is focusable.

pub fn flex(&self) -> f32

Get the flex of the view.

pub fn set_flex(&mut self, flex: f32)

Set the flex of the view.

pub fn is_flex(&self) -> bool

Get whether the view is flexible.

pub fn is_tight(&self) -> bool

Get whether the view is tight.

pub fn set_tight(&mut self, tight: bool)

Set whether the view is tight.

pub fn size(&self) -> Size

Get the size of the view.

pub fn rect(&self) -> Rect

Get the rect of the view in local coordinates.

pub fn transform(&self) -> Affine

Get the transform of the view.

pub fn set_transform(&mut self, transform: Affine)

Set the transform of the view.

pub fn translate(&mut self, translation: Vector)

Translate the transform of the view.

pub fn request_layout(&mut self)

Request a layout of the view tree.

pub fn request_draw(&mut self)

Request a draw of the view tree.

pub fn needs_layout(&self) -> bool

Get whether the view needs to be laid out.

pub fn needs_draw(&self) -> bool

Get whether the view needs to be drawn.

pub fn mark_layed_out(&mut self)

Mark the view as laid out.

This will remove the Update::LAYOUT flag.

pub fn mark_drawn(&mut self)

Mark the view as drawn.

This will remove the Update::DRAW flag.

pub fn update(&self) -> Update

Get the Update of the view.

pub fn cursor(&self) -> Option<Cursor>

Get the cursor of the view.

pub fn set_cursor(&mut self, cursor: Option<Cursor>)

Set the cursor of the view.