ori::core::view
Trait View

pub trait View<T = ()>
where
    T: ?Sized,
{
    type State;

    // Required methods
    fn build(&mut self, cx: &mut BuildCx<'_, '_>, data: &mut T) -> Self::State;
    fn rebuild(
        &mut self,
        state: &mut Self::State,
        cx: &mut RebuildCx<'_, '_>,
        data: &mut T,
        old: &Self,
    );
    fn event(
        &mut self,
        state: &mut Self::State,
        cx: &mut EventCx<'_, '_>,
        data: &mut T,
        event: &Event,
    ) -> bool;
    fn layout(
        &mut self,
        state: &mut Self::State,
        cx: &mut LayoutCx<'_, '_>,
        data: &mut T,
        space: Space,
    ) -> Size;
    fn draw(
        &mut self,
        state: &mut Self::State,
        cx: &mut DrawCx<'_, '_>,
        data: &mut T,
    );
}

A single UI component.

This trait is implemented by all UI components. The user interface is built by composing these components into a view-tree. This operation should be fast, as it is performed very often.

A view also has an associated state type, that is persistent across view-trees. When calling View::build, the view will build it’s state. A view containing another view must also store it’s child’s state. This is usually done by wrapping it in a tuple (MyState, State).

In case a view contains another view the contents should always be wrapped in either State or SeqState. If this is not done strange issues are very likely to occur.

View has four primary methods:

View::rebuild is called after a new view-tree has been built, on the new tree. The view can then compare itself to the old tree and update it’s state accordingly. When a view differs from the old tree, it should call RebuildCx::layout or RebuildCx::draw when applicable. This can be quite tedius to write out, so the Rebuild derive macro can be used to generate this code.

View::event is called when an event occurs. The should then handle the event and return whether it handled it. Command events can be send using BaseCx::cmd.

View::layout is called when the view needs to be laid out. A leaf view should compute it’s own size in accordance with the given Space, and return it. A container view should pass an appropriate Space to it’s contents and the compute it’s own size based on the contents’ size(s).

View::draw is called when the view needs to be drawn.

For examples see the implementation of views like Button or Checkbox.

Required Associated Types

type State

The state of the view, see top-level documentation for more information.

Required Methods

fn build(&mut self, cx: &mut BuildCx<'_, '_>, data: &mut T) -> Self::State

Build the view state, see top-level documentation for more information.

fn rebuild( &mut self, state: &mut Self::State, cx: &mut RebuildCx<'_, '_>, data: &mut T, old: &Self, )

Rebuild the view state, see top-level documentation for more information.

fn event( &mut self, state: &mut Self::State, cx: &mut EventCx<'_, '_>, data: &mut T, event: &Event, ) -> bool

Handle an event, see top-level documentation for more information.

fn layout( &mut self, state: &mut Self::State, cx: &mut LayoutCx<'_, '_>, data: &mut T, space: Space, ) -> Size

Layout the view, see top-level documentation for more information.

fn draw( &mut self, state: &mut Self::State, cx: &mut DrawCx<'_, '_>, data: &mut T, )

Draw the view, see top-level documentation for more information.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.