Struct ori::core::views::LayoutBuilder

pub struct LayoutBuilder<T, V> {
    pub content: Box<dyn FnMut(&mut LayoutCx<'_, '_>, &mut T, Space) -> V>,
}

A view that builds its content based on the layout constraints.

Note that the content is only built on layout.

Fields
content: Box<dyn FnMut(&mut LayoutCx<'_, '_>, &mut T, Space) -> V>

The builder function.

Implementations

impl<T, V> LayoutBuilder<T, V>

pub fn new( builder: impl FnMut(&mut LayoutCx<'_, '_>, &mut T, Space) -> V + 'static ) -> LayoutBuilder<T, V>

Create a new LayoutBuilder view.