ori::core::views
Struct Stack

pub struct Stack<V> {
    pub content: PodSeq<V>,
    pub axis: Axis,
    pub justify: Styled<Justify>,
    pub align: Styled<Align>,
    pub gap: Styled<f32>,
}

A view that stacks it’s content in a line.

Fields
content: PodSeq<V>

The content of the stack.

axis: Axis

The axis of the stack.

justify: Styled<Justify>

How to justify the content along the main axis.

align: Styled<Align>

How to align the content along the cross axis, within each line.

gap: Styled<f32>

The gap between children.

Implementations

impl<V> Stack<V>

pub fn axis(self, axis: impl Into<Axis>) -> Stack<V>

Set self.axis.

The axis of the stack.

pub fn justify(self, justify: impl Into<Styled<Justify>>) -> Stack<V>

Set self.justify.

How to justify the content along the main axis.

pub fn align(self, align: impl Into<Styled<Align>>) -> Stack<V>

Set self.align.

How to align the content along the cross axis, within each line.

pub fn gap(self, gap: impl Into<Styled<f32>>) -> Stack<V>

Set self.gap.

The gap between children.

impl<V> Stack<V>

pub fn new(axis: Axis, content: V) -> Stack<V>

Create a new Stack.

pub fn horizontal(content: V) -> Stack<V>

Create a new horizontal Stack.

pub fn vertical(content: V) -> Stack<V>

Create a new vertical Stack.

impl<T> Stack<Vec<T>>

pub fn vec(axis: Axis) -> Stack<Vec<T>>

Create a new Stack, with vector content.

pub fn vec_horizontal() -> Stack<Vec<T>>

Create a new horizontal Stack, with vector content.

pub fn vec_vertical() -> Stack<Vec<T>>

Create a new vertical Stack, with vector content.

pub fn push(&mut self, view: T)

Push a view to the stack.

pub fn with(self, view: T) -> Stack<Vec<T>>

Push a view to the stack.

pub fn is_empty(&self) -> bool

Get whether the stack is empty.

pub fn len(&self) -> usize

Get the number of views in the stack.

impl<T> Stack<Vec<Box<dyn AnyView<T> + '_>>>

pub fn any(axis: Axis) -> Stack<Vec<Box<dyn AnyView<T> + '_>>>

Create a new Stack, with dynamic content.

pub fn any_horizontal() -> Stack<Vec<Box<dyn AnyView<T> + '_>>>

Create a new horizontal Stack, with dynamic content.

pub fn any_vertical() -> Stack<Vec<Box<dyn AnyView<T> + '_>>>

Create a new vertical Stack, with dynamic content.