Struct ori::core::views::TextInput

pub struct TextInput<T> {

    pub text: Option<String>,
    pub on_change: Option<Box<dyn FnMut(&mut EventCx<'_, '_>, &mut T, String)>>,
    pub on_submit: Option<Box<dyn FnMut(&mut EventCx<'_, '_>, &mut T, String)>>,
    pub placeholder: String,
    pub multiline: bool,
    pub font_size: f32,
    pub font_family: FontFamily,
    pub font_weight: FontWeight,
    pub font_stretch: FontStretch,
    pub font_style: FontStyle,
    pub color: Color,
    pub align: TextAlign,
    pub line_height: f32,
    pub wrap: TextWrap,

}

A text input.

Fields
text: Option<String>

The text.

on_change: Option<Box<dyn FnMut(&mut EventCx<'_, '_>, &mut T, String)>>

A function that returns the text to display.

on_submit: Option<Box<dyn FnMut(&mut EventCx<'_, '_>, &mut T, String)>>

A function that is called when the input is submitted.

placeholder: String

Placeholder text to display when the input is empty.

multiline: bool

Whether the input is multi-line.

When disabled (the default), the input will only accept a single line of text.

font_size: f32

The font size of the text.

font_family: FontFamily

The font family of the text.

font_weight: FontWeight

The font weight of the text.

font_stretch: FontStretch

The font stretch of the text.

font_style: FontStyle

The font.into of the text.

color: Color

The color of the text.

align: TextAlign

The vertical alignment of the text.

line_height: f32

The line height of the text.

wrap: TextWrap

The text wrap of the text.

Implementations

impl<T> TextInput<T>

pub fn placeholder(self, placeholder: impl Into<String>) -> TextInput<T>

Set self.placeholder.

Placeholder text to display when the input is empty.

pub fn multiline(self, multiline: impl Into<bool>) -> TextInput<T>

Set self.multiline.

Whether the input is multi-line.

When disabled (the default), the input will only accept a single line of text.

pub fn font_size(self, font_size: impl Into<f32>) -> TextInput<T>

Set self.font_size.

The font size of the text.

pub fn font_family(self, font_family: impl Into<FontFamily>) -> TextInput<T>

Set self.font_family.

The font family of the text.

pub fn font_weight(self, font_weight: impl Into<FontWeight>) -> TextInput<T>

Set self.font_weight.

The font weight of the text.

pub fn font_stretch(self, font_stretch: impl Into<FontStretch>) -> TextInput<T>

Set self.font_stretch.

The font stretch of the text.

pub fn font_style(self, font_style: impl Into<FontStyle>) -> TextInput<T>

Set self.font_style.

The font.into of the text.

pub fn color(self, color: impl Into<Color>) -> TextInput<T>

Set self.color.

The color of the text.

pub fn align(self, align: impl Into<TextAlign>) -> TextInput<T>

Set self.align.

The vertical alignment of the text.

pub fn line_height(self, line_height: impl Into<f32>) -> TextInput<T>

Set self.line_height.

The line height of the text.

pub fn wrap(self, wrap: impl Into<TextWrap>) -> TextInput<T>

Set self.wrap.

The text wrap of the text.

impl<T> TextInput<T>

pub fn new() -> TextInput<T>

Create a new text input view.

pub fn text(self, text: impl AsRef<str>) -> TextInput<T>

Set the text of the input.

pub fn on_change( self, on_change: impl FnMut(&mut EventCx<'_, '_>, &mut T, String) + 'static ) -> TextInput<T>

Set the callback that is called when the input changes.

Note that this doesn’t trigger a rebuild automatically.

pub fn on_submit( self, on_submit: impl FnMut(&mut EventCx<'_, '_>, &mut T, String) + 'static ) -> TextInput<T>

Set the callback that is called when the input is submitted.