ori
Module prelude

Convenient imports for Ori.

Modules

style

Styling and theming.

Macros

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

hstack

Create a horizontal Stack.

hwrap

Create a horizontal Wrap.

include_font

Load a font from a file or directory.

include_style

Include a style file.

info

Constructs an event at the info level.

is_desktop

Check if the target platform is desktop.

is_mobile

Check if the target platform is mobile.

is_web

Check if the target platform is web.

text

Create a formatted Text.

trace

Constructs an event at the trace level.

vstack

Create a vertical Stack.

vwrap

Create a vertical Wrap.

warn

Constructs an event at the warn level.

zstack

Create a new ZStack view.

Structs

Affine

An affine transformation in 2 dimensional space.

Aligned

A view that aligns its content.

Alignment

Alignment of content inside a container.

Animate

A view that animates.

App

The main application state.

AppBuilder

A builder for an App.

Aspect

A view that lays out its content with a fixed aspect ratio.

BaseCx

A base context that is shared between all other contexts.

BorderRadius

Radi of the corners on a rounded rectangle.

BorderWidth

The border width of a rounded rectangle.

BuildCx

A context for building the view tree.

BuildHandler

A view that hooks into the build cycle.

Button

A button.

ButtonStyle

The derived style for Button.

Canvas

A canvas that can be drawn on.

Checkbox

A checkbox.

CheckboxStyle

The derived style for Checkbox.

Class

A view styled as a class.

Clickable

A click handler.

Clipboard

A clipboard.

Collapsing

A collapsing view.

CollapsingStyle

The derived style for Collapsing.

Color

A color with red, green, blue and alpha components.

ColorPicker

A color picker.

ColorPickerStyle

The derived style for ColorPicker.

CommandProxy

A clonable channel for sending Commands.

Constrain

A view that constrains its content to a given space.

Container

A container view.

ContainerStyle

The derived style for Container.

Curve

A bezier curve.

Decorate

A view for creating custom window decorations.

DelegateCx

The context passed to the [Delegate] trait.

DrawCx

A context for drawing the view tree.

DrawHandler

A view that hooks into the draw cycle.

EventCx

A context for handling events.

EventHandler

A view that handles events.

Flex

The flex value of a view.

Flexible

A flexible view.

Focus

A view that focuses on a part of the data.

FontAttributes

Attributes of a section of text.

FontWeight

A font weight.

Image

An clonable image.

ImageData

Image data.

ImageId

A unique identifier for an Image.

KeyPressed

An event fired when a key is pressed.

Layout

A view for creating views based on the available space.

LayoutCx

A context for laying out the view tree.

Matrix

A 2x2 matrix.

Memo

A view that only builds the inner view when certain data changes.

Modifiers

Modifiers for keyboard and pointer events.

Opaque

A view that bridges the gap between impl View and impl View<T>.

Pad

A view that adds padding to its content.

PadStyle

The derived style for Pad.

Padding

A padding of a rectangle.

Paint

A paint that can be used to fill or stroke a shape.

Painter

A view that draws something.

Paragraph

A paragraph of rich text, that can contain multiple segments with different [TextAttributes].

Pattern

A pattern that can be used to fill a shape.

Pod

A view that has separate ViewState from its content.

PodSeq

Contents of a view, in a sequence.

Point

A point in 2D space.

Pointer

The state of a pointer.

PointerId

A unique pointer id.

PointerMoved

A pointer was moved.

PointerPressed

A pointer button was pressed.

PointerReleased

A pointer button was released.

PointerScrolled

A pointer wheel was scrolled.

RebuildCx

A context for rebuilding the view tree.

RebuildHandler

A view that handles rebuilds.

Rect

A rectangle defined by its minimum and maximum points.

Scroll

A scrollable view.

ScrollStyle

The derived style for Scroll.

SeqState

The state of a PodSeq.

Size

A 2 dimensional size.

Slider

A slider.

SliderStyle

The derived style for Slider.

Space

Space available to lay out a view.

Stack

A view that stacks it’s content in a line.

StackStyle

The derived style for Stack.

State

The state of a Pod.

Stroke

Properties of a stroke.

Style

A style.

Styled

A styled value.

Styles

A collection of styles.

Suspense

A view that suspends rendering while a future is pending.

Text

A view that displays text.

TextInput

A text input.

TextInputStyle

The derived style for TextInput.

TextStyle

The derived style for Text.

Theme

A theme.

Tooltip

A view that displays some text when the content is hovered.

TooltipStyle

The derived style for Tooltip.

Transform

A view that transforms its content.

Transition

A transition.

Trigger

A view that creates a trigger around the content.

Vector

A 2D vector.

ViewState

State associated with a View.

Window

A window.

WindowCloseRequested

Event emitted when a window wants to close.

WindowId

A unique identifier for a window.

WithState

A view that stores some additional data.

WithStyle

A view that applies a style to its content.

Wrap

A view that lays out it’s content in a line wrapping if it doesn’t fit.

WrapStyle

The derived style for Wrap.

ZStack

A view that overlays its content on top of each other.

Enums

Align

The alignment of items along the cross axis.

AppCommand

Commands that can be sent to the application.

Axis

An axis is a direction in which a layout is applied.

BlendMode

Ways to blend two colors.

ClickEvent

A click event.

Code

A keyboard key-code.

Cursor

A mouse cursor icon.

Easing

A transition easing curve.

Event

An event that can be sent to a view.

FillRule

Rule determining if a point is inside a shape.

FontFamily

A font family, by default FontFamily::SansSerif.

FontSource

A source for a font.

FontStretch

A font stretch.

FontStyle

A font style.

Justify

The justify content of a stack container.

Key

A keyboard key.

PointerButton

A pointer button.

Shader

Ways to fill a shape.

StrokeCap

Ways to draw the end of a stroke.

StrokeJoin

Ways to join two segments of a stroke.

TextAlign

Alignment of a section of text.

TextWrap

Wrapping of a section of text.

WindowSizing

The sizing of a window.

Constants

FILL

A constant used to indicate that a dimension should fill the available space.

Traits

AnyView

A view that supports dynamic dispatch.

AppDelegate

A delegate for handling events in an application.

Fonts

A trait for managing fonts and text layout.

Rebuild

A convenience trait for rebuilding a View.

View

A single UI component.

ViewSeq

A sequence of views.

Functions

align

Create a new Aligned view.

animate

Create a new Animate.

any

Create a new BoxedView.

aspect

Create a new Aspect view.

background

Create a new Container with background.

bottom

Create a new Aligned view that aligns its content to the bottom.

bottom_left

Create a new Aligned view that aligns its content to the bottom left.

bottom_right

Create a new Aligned view that aligns its content to the bottom right.

button

Create a new Button.

center

Create a new Aligned view that aligns its content to the center.

checkbox

Create a new Checkbox.

circle

Create a new Painter view that draws a circle.

class

Wrap a view in a class.

collapsing

Create a new Collapsing.

color_picker

Create a new ColorPicker.

comp

Create a computed style, shorthand for Styled::computed.

constrain

Create a new Constrained view, constraining its content to a space.

container

Create a new Container.

decorate

A view for creating custom window decorations.

ease

Create an ease transition with the given duration.

ellipse

Create a new Painter view that draws an ellipse.

expand

Create a new expanded Flexible view.

flex

Create a new Flexible view.

focus

Create a new Focus.

height

Create a new Constrained view, constraining its content to a height.

hex

Create a new color, with the given hex string.

hscroll

Create a new horizontal Scroll.

hsl

Create a new color, with the given hue, saturation and lightness components.

hsla

Create a new color, with the given hue, saturation, lightness and alpha components.

hstack

Create a horizontal Stack.

hstack_any

Create a horizontal Stack, with dynamic content.

hstack_vec

Create a horizontal Stack, with vector content.

hsv

Create a new color, with the given hue, saturation and value components.

hsva

Create a new color, with the given hue, saturation, value and alpha components.

hwrap

Create a horizontal Wrap.

hwrap_any

Create a horizontal Wrap, with dynamic content.

hwrap_vec

Create a horizontal Wrap, with a vector of content.

layout

A view for creating views based on the available space.

left

Create a new Aligned view that aligns its content to the left.

linear

Create a linear transition with the given duration.

max_height

Create a new Constrained view, constraining its content to a maximum height.

max_size

Create a new Constrained view, constraining its content to a maximum size.

max_width

Create a new Constrained view, constraining its content to a maximum width.

memo

Create a new Memo.

min_height

Create a new Constrained view, constraining its content to a minimum height.

min_size

Create a new Constrained view, constraining its content to a minimum size.

min_width

Create a new Constrained view, constraining its content to a minimum width.

okhsl

Create a new color, with the given hue, saturation and lightness components.

okhsla

Create a new color, with the given hue, saturation, lightness and alpha components.

okhsv

Create a new color, with the given hue, saturation and value components.

okhsva

Create a new color, with the given hue, saturation, value and alpha components.

oklab

Create a new color, with the given lightness, a and b components.

oklaba

Create a new color, with the given lightness, a, b and alpha components.

oklch

Create a new color, with the given lightness, chroma and hue components.

oklcha

Create a new color, with the given lightness, chroma, hue and alpha components.

on_build

Create a new BuildHandler.

on_click

Create a new Clickable, that calls on_click when clicked.

on_draw

Create a new DrawHandler.

on_event

Create a new EventHandler, with an after callback.

on_event_before

Create a new EventHandler, with a before callback.

on_press

Create a new Clickable, that calls on_press when pressed.

on_rebuild

Create a new RebuildHandler.

on_release

Create a new Clickable, that calls on_release when released.

opaque

A view that bridges the gap between impl View and impl View<T>.

pad

Create a new Pad view.

pad_bottom

Create a new Pad view adding padding to the bottom.

pad_left

Create a new Pad view adding padding to the left.

pad_right

Create a new Pad view adding padding to the right.

pad_top

Create a new Pad view adding padding to the top.

padded

Create a new Pad view with padding from the padding style tag.

painter

Create a new Painter view.

pod

Create a new Pod view.

pt

Convert a from points to pixels.

rect

Create a new Painter view that draws a rectangle.

rgb

Create a new color, with the given red, green and blue components.

rgba

Create a new color, with the given red, green, blue and alpha components.

right

Create a new Aligned view that aligns its content to the right.

rotate

Create a new Transform view that rotates its content.

rotate_degrees

Create a new Transform view that rotates its content in degrees.

scale

Create a new Transform view that scales its content.

size

Create a new Constrained view, cosntraining its content to a size.

slider

Create a new Slider.

style

Create a style key, shorthand for Styled::style.

suspense

Create a new Suspense view.

text

Create a new Text.

text_input

Create a new TextInput.

tooltip

Create a new Tooltip view.

top

Create a new Aligned view that aligns its content to the top.

top_left

Create a new Aligned view that aligns its content to the top left.

top_right

Create a new Aligned view that aligns its content to the top right.

transform

Create a new Transform view.

transition

Animate a transition.

transition_active

Animate a view when active changes.

transition_focused

Animate a view when focused changes.

transition_hovered

Animate a view when hovered changes.

translate

Create a new Transform view that translates its content.

trigger

Create a new Trigger view.

val

Create a style value, shorthand for Styled::new.

vscroll

Create a new vertical Scroll.

vstack

Create a vertical Stack.

vstack_any

Create a vertical Stack, with dynamic content.

vstack_vec

Create a vertical Stack, with vector content.

vwrap

Create a vertical Wrap.

vwrap_any

Create a vertical Wrap, with dynamic content.

vwrap_vec

Create a vertical Wrap, with a vector of content.

width

Create a new Constrained view, constraining its content to a width.

with_data

Create a new WithState that replaces the data with the state.

with_data_default

Create a new WithState that replaces the data with the state using S::default().

with_state

Create a new WithState.

with_state_default

Create a new WithState using S::default().

with_style

Create a view that applies a style to its content.

without_data

Create a new view unwrapping some data from the state.

without_state

Create a new view unwrapping some state from the data.

zstack

Create a new ZStack view.

Type Aliases

BoxedView

A boxed dynamic view.

Lens

A lens used by Focus.

Attribute Macros

desktop

Only include the annotated item on desktop platforms.

mobile

Only include the annotated item on mobile platforms.

web

Only include the annotated item on web platforms.

Derive Macros

Build

Derive the builder pattern.

Rebuild

Derive the Rebuild trait.

Styled

Derived the styled pattern.