ori::prelude
Enum Event

#[non_exhaustive]
pub enum Event {

    WindowResized(WindowResized),
    WindowScaled(WindowScaled),
    WindowMaximized(WindowMaximized),
    WindowCloseRequested(WindowCloseRequested),
    PointerMoved(PointerMoved),
    PointerLeft(PointerLeft),
    PointerPressed(PointerPressed),
    PointerReleased(PointerReleased),
    PointerScrolled(PointerScrolled),
    KeyPressed(KeyPressed),
    KeyReleased(KeyReleased),
    FocusNext,
    FocusPrev,
    FocusWanted,
    FocusGiven(FocusTarget),
    Animate(f32),
    Command(Command),
    Notify,

}

An event that can be sent to a view.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

WindowResized(WindowResized)

The window was resized.

WindowScaled(WindowScaled)

The window was scaled.

WindowMaximized(WindowMaximized)

The window was maximized.

WindowCloseRequested(WindowCloseRequested)

The window requested to be close.

PointerMoved(PointerMoved)

A pointer moved.

PointerLeft(PointerLeft)

A pointer left the window.

PointerPressed(PointerPressed)

A pointer button was pressed.

PointerReleased(PointerReleased)

A pointer button was released.

PointerScrolled(PointerScrolled)

A pointer was scrolled.

KeyPressed(KeyPressed)

A keyboard key was pressed.

KeyReleased(KeyReleased)

A keyboard key was released.

FocusNext

Focus should be switched to next view in the focus chain.

FocusPrev

Focus should be switched to previous view in the focus chain.

FocusWanted

Focus is wanted by another view.

A view receiving this event should give up focus.

FocusGiven(FocusTarget)

Focus given to either a specific view or any focu

Animate(f32)

An animation frame has passed.

Command(Command)

A command was sent.

Notify

Event sent when something has changed and the view should be given a chance to update.

Implementations

impl Event

pub fn is_cmd<T>(&self) -> bool
where T: Any,

Check if the event is a command of a specific type.

pub fn cmd<T>(&self) -> Option<&T>
where T: Any,

Try to get the command as a specific type.

Returns None if the event is not a command or if the command is not of the specified type.

pub fn is_key_pressed(&self, key: impl IsKey) -> bool

Check if the event represents a key press of a specific key.

pub fn is_key_released(&self, key: impl IsKey) -> bool

Check if the event represents a key release of a specific key.

pub fn wants_focus(&self) -> bool

Check if the event wants to take focus.

This is true for FocusNext, FocusPrev, and FocusWanted.