ori::prelude
Struct DrawCx

pub struct DrawCx<'a, 'b> { /* private fields */ }

A context for drawing the view tree.

Implementations

impl<'a, 'b> DrawCx<'a, 'b>

pub fn new( base: &'a mut BaseCx<'b>, view_state: &'a mut ViewState, canvas: &'a mut Canvas, ) -> DrawCx<'a, 'b>

Create a new draw context.

pub fn child(&mut self) -> DrawCx<'_, 'b>

Create a child context.

pub fn is_visible(&self, rect: Rect) -> bool

Check if a rect is visible.

pub fn transform(&self) -> Affine

Get the transform of the view.

pub fn size(&self) -> Size

Get the size of the view.

pub fn rect(&self) -> Rect

Get the rect of the view in local space.

pub fn canvas(&mut self) -> &mut Canvas

Get the canvas.

pub fn fill_rect(&mut self, rect: Rect, paint: impl Into<Paint>)

Draw a rectangle.

pub fn trigger(&mut self, rect: Rect)

Draw a trigger rectangle.

pub fn fill(&mut self, curve: Curve, fill: FillRule, paint: impl Into<Paint>)

Fill a curve.

pub fn stroke( &mut self, curve: Curve, stroke: impl Into<Stroke>, paint: impl Into<Paint>, )

Stroke a curve.

pub fn text(&mut self, text: impl Display, rect: Rect, font: FontAttributes)

Draw some text.

pub fn paragraph(&mut self, paragraph: &Paragraph, rect: Rect)

Draw a paragraph.

pub fn quad( &mut self, rect: Rect, paint: impl Into<Paint>, border_radius: impl Into<BorderRadius>, border_width: impl Into<BorderWidth>, border_paint: impl Into<Paint>, )

Draw a rectangle with rounded corners and a border.

pub fn draw_canvas(&mut self, canvas: Canvas)

Draw a canvas.

pub fn overlay<T>( &mut self, index: i32, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T, ) -> T

Draw an overlay, at index.

pub fn layer<T>( &mut self, transform: Affine, mask: Option<Mask>, view: Option<ViewId>, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T, ) -> T

Draw a layer, with a transform, mask, and view.

pub fn hoverable<T>(&mut self, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T) -> T

Draw a hoverable layer.

pub fn transformed<T>( &mut self, transform: Affine, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T, ) -> T

Draw a layer with a transform.

pub fn translated<T>( &mut self, translation: Vector, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T, ) -> T

Draw a layer with a translation.

pub fn rotated<T>( &mut self, angle: f32, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T, ) -> T

Draw a layer with a rotation.

pub fn scaled<T>( &mut self, scale: Vector, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T, ) -> T

Draw a layer with a scale.

pub fn masked<T>( &mut self, mask: impl Into<Mask>, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T, ) -> T

Draw a layer with a mask.

impl DrawCx<'_, '_>

pub fn window(&self) -> &Window

Get the window.

pub fn window_mut(&mut self) -> &mut Window

Get the window mutably.

pub fn styles(&self) -> &Styles

Get the styles.

pub fn id(&self) -> ViewId

Get the id of the view.

pub fn is_hovered(&self) -> bool

Get whether the view is hovered.

pub fn is_focused(&self) -> bool

Get whether the view is focused.

pub fn is_active(&self) -> bool

Get whether the view is active.

pub fn is_focusable(&self) -> bool

Get whether the view is focusable.

pub fn has_hovered(&self) -> bool

Get whether a child view is hovered.

pub fn has_focused(&self) -> bool

Get whether a child view is focused.

pub fn has_active(&self) -> bool

Get whether a child view is active.

pub fn contains_property<T>(&self) -> bool
where T: 'static,

Check if the view has the property T.

pub fn insert_property<T>(&mut self, item: T)
where T: 'static,

Insert a property into the view.

pub fn remove_property<T>(&mut self) -> Option<T>
where T: 'static,

Remove a property from the view.

pub fn get_property<T>(&self) -> Option<&T>
where T: 'static,

Get the property T of the view.

pub fn get_property_mut<T>(&mut self) -> Option<&mut T>
where T: 'static,

Get the property T of the view mutably.

pub fn property_or_insert_with<T, F>(&mut self, f: F) -> &mut T
where T: 'static, F: FnOnce() -> T,

Get the property T of the view or insert it with a value.

pub fn property_or<T>(&mut self, item: T) -> &mut T
where T: 'static,

Get the property T of the view or insert it with a value.

pub fn property_or_default<T>(&mut self) -> &mut T
where T: 'static + Default,

Get the property T of the view or insert it with a default value.

pub fn class(&self) -> Option<&str>

Get the class of the view.