Struct ori::prelude::EventCx

pub struct EventCx<'a, 'b> { /* private fields */ }

A context for handling events.

Implementations

impl<'a, 'b> EventCx<'a, 'b>

pub fn new( base: &'a mut BaseCx<'b>, view_state: &'a mut ViewState, rebuild: &'a mut bool, ) -> EventCx<'a, 'b>

Create a new event context.

pub fn child(&mut self) -> EventCx<'_, 'b>

Create a child context.

pub fn as_build_cx(&mut self) -> BuildCx<'_, 'b>

Get a build context.

pub fn as_rebuild_cx(&mut self) -> RebuildCx<'_, 'b>

Get a rebuild context.

pub fn size(&self) -> Size

Get the size of the view.

pub fn rect(&self) -> Rect

Get the rect of the view in local space.

pub fn transform(&self) -> Affine

Get the transform of the view.

pub fn local(&self, point: Point) -> Point

Transform a point from global space to local space.

pub fn rebuild(&mut self)

Request a rebuild of the view tree.

pub fn layout(&mut self)

Request a layout of the view tree.

pub fn draw(&mut self)

Request a draw of the view tree.

pub fn animate(&mut self)

Request an animation frame.

pub fn set_cursor(&mut self, cursor: Option<Cursor>)

Set the cursor of the view.

pub fn set_hot(&mut self, hot: bool) -> bool

Set whether the view is hot.

Returns true if the hot state changed.

pub fn set_focused(&mut self, focused: bool) -> bool

Set whether the view is focused.

Returns true if the focused state changed.

pub fn set_active(&mut self, active: bool) -> bool

Set whether the view is active.

Returns true if the active state changed.

pub fn was_hot(&self) -> bool

Get whether the view was hot last call.

pub fn was_focused(&self) -> bool

Get whether the view was focused last call.

pub fn was_active(&self) -> bool

Get whether the view was active last call.

pub fn had_hot(&self) -> bool

Get whether a child view was hot last call.

pub fn had_focused(&self) -> bool

Get whether a child view was focused last call.

pub fn had_active(&self) -> bool

Get whether a child view was active last call.

pub fn hot_changed(&self) -> bool

Get whether the view’s hot state changed.

pub fn focused_changed(&self) -> bool

Get whether the view’s focused state changed.

pub fn active_changed(&self) -> bool

Get whether the view’s active state changed.

pub fn has_hot_changed(&self) -> bool

Get whether a child view’s hot state changed.

pub fn has_focused_changed(&self) -> bool

Get whether a child view’s focused state changed.

pub fn has_active_changed(&self) -> bool

Get whether a child view’s active state changed.

impl EventCx<'_, '_>

pub fn window(&mut self) -> &mut Window

Get the window.

pub fn id(&self) -> ViewId

Get the id of the view.

pub fn is_hot(&self) -> bool

Get whether the view is hot.

pub fn is_focused(&self) -> bool

Get whether the view is focused.

pub fn is_active(&self) -> bool

Get whether the view is active.

pub fn has_hot(&self) -> bool

Get whether a child view is hot.

pub fn has_focused(&self) -> bool

Get whether a child view is focused.

pub fn has_active(&self) -> bool

Get whether a child view is active.

pub fn contains_property<T>(&self) -> bool
where T: 'static,

Check if the view has the property T.

pub fn insert_property<T>(&mut self, item: T)
where T: 'static,

Insert a property into the view.

pub fn remove_property<T>(&mut self) -> Option<T>
where T: 'static,

Remove a property from the view.

pub fn get_property<T>(&self) -> Option<&T>
where T: 'static,

Get the property T of the view.

pub fn get_property_mut<T>(&mut self) -> Option<&mut T>
where T: 'static,

Get the property T of the view mutably.

pub fn property_or_insert_with<T, F>(&mut self, f: F) -> &mut T
where T: 'static, F: FnOnce() -> T,

Get the property T of the view or insert it with a value.

pub fn property_or<T>(&mut self, item: T) -> &mut T
where T: 'static,

Get the property T of the view or insert it with a value.

pub fn property_or_default<T>(&mut self) -> &mut T
where T: 'static + Default,

Get the property T of the view or insert it with a default value.