ori::prelude
Struct Point

#[repr(C)]
pub struct Point {
    pub x: f32,
    pub y: f32,
}

A point in 2D space.

Fields
x: f32

The x coordinate.

y: f32

The y coordinate.

Implementations

impl Point

pub const ZERO: Point = _

The zero point.

pub const ONE: Point = _

The one point.

pub const X: Point = _

The unit x point.

pub const Y: Point = _

The unit y point.

pub const NEG_X: Point = _

The negative unit x point.

pub const NEG_Y: Point = _

The negative unit y point.

pub const fn new(x: f32, y: f32) -> Point

Create a new point.

pub const fn all(value: f32) -> Point

Create a new point with the same x and y.

pub fn min(self, other: Point) -> Point

Get the min of self and other by element.

pub fn max(self, other: Point) -> Point

Get the max of self and other by element.

pub fn clamp(self, min: Point, max: Point) -> Point

Clamp self to the range [min, max] by element.

pub fn floor(self) -> Point

Floor the point by element.

pub fn ceil(self) -> Point

Ceil the point by element.

pub fn round(self) -> Point

Round the point by element.

pub fn fract(self) -> Point

Get the fractional component by element.

pub fn is_finite(self) -> bool

Check if the point is finite.

pub fn is_infinite(self) -> bool

Check if the point is infinite.

pub fn is_nan(self) -> bool

Check if the point is NaN.

pub fn distance(self, other: Point) -> f32

Compute the dot distance between two points.

pub fn lerp(self, other: Point, t: f32) -> Point

Linearly interpolate between two points.

pub const fn to_vector(self) -> Vector

Convert the point to a vector.

pub const fn to_size(self) -> Size

Convert the point to a size.