ori::prelude
Struct Size

#[repr(C)]
pub struct Size {
    pub width: f32,
    pub height: f32,
}

A 2 dimensional size.

Fields
width: f32

The width.

height: f32

The height.

Implementations

impl Size

pub const ZERO: Size = _

The zero size.

pub const UNBOUNDED: Size = _

The unbounded size.

pub const INFINITY: Size = _

The infinite size.

pub const FILL: Size = Self::INFINITY

Alias for Self::INFINITY.

pub const fn new(width: f32, height: f32) -> Size

Create a new size.

pub const fn all(value: f32) -> Size

Create a new size with the same width and height.

pub fn min(self, other: Size) -> Size

Get the min of self and other by element.

pub fn min_element(self) -> f32

Get the smallest element of self.

pub fn max(self, other: Size) -> Size

Get the max of self and other by element.

pub fn max_element(self) -> f32

Get the largest element of self.

pub fn clamp(self, min: Size, max: Size) -> Size

Clamp self to the range [min, max] by element.

pub fn floor(self) -> Size

Floor self by element.

pub fn ceil(self) -> Size

Ceil self by element.

pub fn round(self) -> Size

Round self by element.

pub fn finite_or_zero(self) -> Size

If self is finite, return self, otherwise return zero. Applied by element.

pub fn is_finite(self) -> bool

Get whether the size is finite.

pub fn is_infinite(self) -> bool

Get whether the size is infinite.

pub const fn to_point(self) -> Point

Convert the size to a vector.

pub const fn to_vector(self) -> Vector

Convert the size to a vector.