Struct ori::prelude::Space

pub struct Space {
    pub min: Size,
    pub max: Size,
}

Space available to lay out a view.

Fields
min: Size

Minimum size the view can be.

max: Size

Maximum size the view can be.

Implementations

impl Space

pub const ZERO: Space = _

The zero space.

pub const UNBOUNDED: Space = _

The unbounded space.

pub const FILL: Space = _

The infinite space.

pub const fn new(min: Size, max: Size) -> Space

Create a new space.

pub fn from_size(size: Size) -> Space

Create a new space with the same minimum and maximum size.

pub fn shrink(self, size: Size) -> Space

Shrink the space by size.

pub fn expand(self, size: Size) -> Space

Expand the space by size.

pub fn loosen(self) -> Space

Loosen the space, setting the minimum size to zero.

pub fn loosen_width(self) -> Space

Loosen the width, setting the minimum width to zero.

pub fn loosen_height(self) -> Space

Loosen the height, setting the minimum height to zero.

pub fn constrain(self, other: Space) -> Space

Get the most constraning space between self and `other

pub fn fit(self, size: Size) -> Size

Clamp a size to the space.

pub fn is_finite(self) -> bool

Get whether the space is finite.

pub fn is_infinite(self) -> bool

Get whether the space is infinite.