ori::prelude
Struct ViewState

pub struct ViewState { /* private fields */ }

State associated with a View.

Implementations

impl ViewState

pub fn new(id: ViewId) -> ViewState

Create a new ViewState with the given ViewId.

pub fn prepare(&mut self)

Prepare the view.

pub fn propagate(&mut self, child: &mut ViewState)

Propagate the state of a child view.

pub fn id(&self) -> ViewId

Get the id of the view.

pub fn is_hovered(&self) -> bool

Get whether the view is hovered.

pub fn set_hovered(&mut self, hovered: bool)

Set whether the view is hovered.

pub fn is_focused(&self) -> bool

Get whether the view is focused.

pub fn set_focused(&mut self, focused: bool)

Set whether the view is focused.

pub fn is_active(&self) -> bool

Get whether the view is active.

pub fn set_active(&mut self, active: bool)

Set whether the view is active.

pub fn has_hovered(&self) -> bool

Get whether the view has a hovered child.

pub fn has_focused(&self) -> bool

Get whether the view has a focused child.

pub fn has_active(&self) -> bool

Get whether the view has an active child.

pub fn is_focusable(&self) -> bool

Get whether the view is focusable.

pub fn set_focusable(&mut self, focusable: bool)

Set whether the view is focusable.

pub fn contains_property<T>(&self) -> bool
where T: 'static,

Check if the view has the property T.

pub fn insert_property<T>(&mut self, item: T)
where T: 'static,

Insert a property into the view.

pub fn remove_property<T>(&mut self) -> Option<T>
where T: 'static,

Remove a property from the view.

pub fn get_property<T>(&self) -> Option<&T>
where T: 'static,

Get the property T of the view.

pub fn get_property_mut<T>(&mut self) -> Option<&mut T>
where T: 'static,

Get the property T of the view mutably.

pub fn property_or_insert_with<T, F>(&mut self, f: F) -> &mut T
where T: 'static, F: FnOnce() -> T,

Get the property T of the view or insert it with a value.

pub fn property_or<T>(&mut self, item: T) -> &mut T
where T: 'static,

Get the property T of the view or insert it with a value.

pub fn property_or_default<T>(&mut self) -> &mut T
where T: 'static + Default,

Get the property T of the view or insert it with a default value.

pub fn set_size(&mut self, size: Size)

Set the size of the view.

pub fn size(&self) -> Size

Get the size of the view.

pub fn rect(&self) -> Rect

Get the rect of the view in local coordinates.

pub fn transform(&self) -> Affine

Get the transform of the view.

pub fn set_transform(&mut self, transform: Affine)

Set the transform of the view.

pub fn translate(&mut self, translation: Vector)

Translate the transform of the view.

pub fn request_layout(&mut self)

Request a layout of the view tree.

pub fn request_draw(&mut self)

Request a draw of the view tree.

pub fn request_animate(&mut self)

Request an animation frame of the view tree.

pub fn needs_layout(&self) -> bool

Get whether the view needs to be laid out.

pub fn needs_draw(&self) -> bool

Get whether the view needs to be drawn.

pub fn needs_animate(&self) -> bool

Get whether the view needs an animation frame.

pub fn mark_layed_out(&mut self)

Mark the view as laid out.

This will remove the Update::LAYOUT flag.

pub fn mark_drawn(&mut self)

Mark the view as drawn.

This will remove the Update::DRAW flag.

pub fn mark_animated(&mut self)

Mark the view as animated.

This will remove the Update::ANIMATE flag.

pub fn flags(&self) -> ViewFlags

Get the flags of the view.

pub fn update(&self) -> Update

Get the Update of the view.

pub fn cursor(&self) -> Option<Cursor>

Get the cursor of the view.

pub fn set_cursor(&mut self, cursor: Option<Cursor>)

Set the cursor of the view.

pub fn ime(&self) -> Option<&Ime>

Get the IME of the view.

pub fn set_ime(&mut self, ime: Option<Ime>)

Set the IME of the view.