ori::prelude::style
Enum Styled

pub enum Styled<T> {
    Value(T),
    Style(Style<T>),
    Computed(Arc<Box<dyn Fn(&Styles) -> T + Sync + Send>>),
}

A styled value.

Variants

Value(T)

A value.

Style(Style<T>)

A style key.

Computed(Arc<Box<dyn Fn(&Styles) -> T + Sync + Send>>)

A derived style.

Note: This is by far the least efficient variant, and should only be used when necessary.

Implementations

impl<T> Styled<T>

pub fn new(style: impl Into<Styled<T>>) -> Styled<T>

Create a new styled value.

pub fn style(key: impl Into<Style<T>>) -> Styled<T>

Create a new styled from a style key.

pub fn get(&self, styles: &Styles) -> Option<T>
where T: Clone + Send + Sync + 'static,

Get the value, or a style from the styles.

pub fn get_or(&self, styles: &Styles, default: T) -> T
where T: Clone + Send + Sync + 'static,

Get the value, or a style from the styles.

pub fn get_or_else<F>(&self, styles: &Styles, default: F) -> T
where T: Clone + Send + Sync + 'static, F: FnOnce() -> T,

Get the value, or a style from the styles.