Module ori::core::canvas

Canvas module.

This module contains the Canvas type, which is used to draw primitives to a Scene.

Structs

Background

A background.

Batch

A batch of meshes to draw.

BorderRadius

Radi of the corners on a Quad.

BorderWidth

The border width of a Quad.

BoxShadow

A box shadow.

Canvas

A canvas used for drawing a Scene.

Color

A color with red, green, blue and alpha components.

Curve

A curve.

DisplayHex

A type for displaying a color as a hex string.

Fragment

A item to be rendered.

Mesh

A mesh containing vertices, indices and an optional image.

Quad

A quad primitive.

Scene

A collection of fragments to be rendered.

Vertex

A vertex in a Mesh.

Enums

Primitive

A primitive to be rendered.

Functions

hex

Create a new color, with the given hex string.

hsl

Create a new color, with the given hue, saturation and lightness components.

hsla

Create a new color, with the given hue, saturation, lightness and alpha components.

oklab

Create a new color, with the given lightness, a and b components.

oklaba

Create a new color, with the given lightness, a, b and alpha components.

rgb

Create a new color, with the given red, green and blue components.

rgba

Create a new color, with the given red, green, blue and alpha components.