Struct ori::core::canvas::Curve

pub struct Curve {
    pub points: Vec<Point>,
}

A curve.

Fields
points: Vec<Point>

The points of the curve.

Implementations

impl Curve

pub const RESOLUTION: f32 = 3f32

The resolution of the curve, measured in pixels per point.

pub const fn new() -> Curve

Create an empty curve.

pub fn extend(&mut self, other: impl IntoIterator<Item = Point>)

Extend the curve with the points from other.

pub fn arc_center_angle( center: Point, radius: f32, start_angle: f32, end_angle: f32 ) -> Curve

Create an arc.

Arguments

center: The center of the arc.

radius: The radius of the arc.

start_angle: The start angle of the arc.

end_angle: The end angle of the arc.

pub fn parametric(f: impl Fn(f32) -> Point, start: f32, end: f32) -> Curve

Create a parametric curve.

Arguments

f: The function that returns the point at a given time.

start: The start time.

end: The end time.

pub fn len(&self) -> usize

Get the number of points in the curve.

pub fn is_empty(&self) -> bool

Get whether the curve is empty.

pub fn push(&mut self, point: Point)

Add a point to the curve.

pub fn remove(&mut self, index: usize)

Remove a point from the curve at index.

pub fn clear(&mut self)

Clear the curve.

pub fn iter(&self) -> impl DoubleEndedIterator

Get an iterator over the points of the curve.

pub fn iter_mut(&mut self) -> impl DoubleEndedIterator

Get a mutable iterator over the points of the curve.

pub fn stroke(&self, thickness: f32, color: Color) -> Mesh

Creates a mesh with rounded ends.