Module ori::log

Ori log module.

Modules

callsite

Callsites represent the source locations from which spans or events originate.

dispatcher

Dispatches trace events to Subscribers.

event

Events represent single points in time during the execution of a program.

field

Span and Event key-value data.

instrument

Attach a span to a std::future::Future.

level_filters

Trace verbosity level filtering.

metadata

Metadata describing trace data.

span

Spans represent periods of time in which a program was executing in a particular context.

subscriber

Collects and records trace data.

Macros

debug

Constructs an event at the debug level.

debug_span

Constructs a span at the debug level.

enabled

Checks whether a span or event is enabled based on the provided metadata.

error

Constructs an event at the error level.

error_span

Constructs a span at the error level.

event

Constructs a new Event.

event_enabled

Tests whether an event with the specified level and target would be enabled.

info

Constructs an event at the info level.

info_span

Constructs a span at the info level.

metadata

Statically constructs new span metadata.

span

Constructs a new span.

span_enabled

Tests whether a span with the specified level and target would be enabled.

trace

Constructs an event at the trace level.

trace_span

Constructs a span at the trace level.

warn

Constructs an event at the warn level.

warn_span

Constructs a span at the warn level.

Structs

Dispatch

Dispatch trace data to a Subscriber.

Event

Events represent single points in time where something occurred during the execution of a program.

Id

Identifies a span within the context of a subscriber.

Level

Describes the level of verbosity of a span or event.

Metadata

Metadata describing a span or event.

Span

A handle representing a span, with the capability to enter the span if it exists.

Traits

Callsite

Trait implemented by callsites.

Instrument

Attaches spans to a std::future::Future.

Subscriber

Trait representing the functions required to collect trace data.

Value

A field value of an erased type.

Attribute Macros

instrument

Instruments a function to create and enter a tracing span every time the function is called.