Struct ori::prelude::DrawCx

pub struct DrawCx<'a, 'b> { /* private fields */ }

A context for drawing the view tree.

Implementations

impl<'a, 'b> DrawCx<'a, 'b>

pub fn new( base: &'a mut BaseCx<'b>, view_state: &'a mut ViewState, window: &'a mut Window, canvas: &'a mut Canvas ) -> DrawCx<'a, 'b>

Create a new draw context.

pub fn child(&mut self) -> DrawCx<'_, 'b>

Create a child context.

pub fn rebuild_cx(&mut self) -> RebuildCx<'_, 'b>

Get a rebuild context.

pub fn transform(&self) -> Affine

Get the transform of the view.

pub fn size(&self) -> Size

Get the size of the view.

pub fn rect(&self) -> Rect

Get the rect of the view in local space.

pub fn canvas(&mut self) -> &mut Canvas

Get the canvas.

pub fn fill_rect(&mut self, rect: Rect, paint: impl Into<Paint>)

Draw a rectangle.

pub fn trigger(&mut self, rect: Rect)

Draw a trigger rectangle.

pub fn fill_curve( &mut self, curve: Curve, fill: FillRule, paint: impl Into<Paint> )

Fill a curve.

pub fn stroke( &mut self, curve: Curve, stroke: impl Into<Stroke>, paint: impl Into<Paint> )

Stroke a curve.

pub fn image(&mut self, point: Point, image: Image)

Draw an image.

pub fn text(&mut self, offset: Vector, buffer: &TextBuffer)

Draw a text buffer.

pub fn text_raw(&mut self, offset: Vector, buffer: &Buffer)

Draw a raw text buffer.

pub fn quad( &mut self, rect: Rect, paint: impl Into<Paint>, border_radius: impl Into<BorderRadius>, border_width: impl Into<BorderWidth>, border_paint: impl Into<Paint> )

Draw a rectangle with rounded corners and a border.

pub fn overlay<T>( &mut self, index: i32, f: impl FnOnce(&mut DrawCx<'_, 'b>) -> T ) -> T

Draw an overlay, at index.

pub fn hoverable(&mut self, f: impl FnOnce(&mut DrawCx<'_, 'b>))

Draw a hoverable layer.

pub fn layer(&mut self, transform: Affine, f: impl FnOnce(&mut DrawCx<'_, 'b>))

Draw a layer.

pub fn translate( &mut self, translation: Vector, f: impl FnOnce(&mut DrawCx<'_, 'b>) )

Draw a layer with a translation.

pub fn rotate(&mut self, angle: f32, f: impl FnOnce(&mut DrawCx<'_, 'b>))

Draw a layer with a rotation.

pub fn mask( &mut self, mask: impl Into<Mask>, f: impl FnOnce(&mut DrawCx<'_, 'b>) )

Draw a layer with a mask.

impl DrawCx<'_, '_>

pub fn window(&mut self) -> &mut Window

Get the window.

pub fn id(&self) -> ViewId

Get the id of the view.

pub fn is_hot(&self) -> bool

Get whether the view is hot.

pub fn is_focused(&self) -> bool

Get whether the view is focused.

pub fn is_active(&self) -> bool

Get whether the view is active.

pub fn has_hot(&self) -> bool

Get whether a child view is hot.

pub fn has_focused(&self) -> bool

Get whether a child view is focused.

pub fn has_active(&self) -> bool

Get whether a child view is active.

pub fn flex(&self) -> f32

Get the flex of the view.

pub fn is_tight(&self) -> bool

Get whether the view is tight.