Struct ori::prelude::Window

pub struct Window {
    pub title: String,
    pub icon: Option<Image>,
    pub size: Size,
    pub sizing: WindowSizing,
    pub scale: f32,
    pub resizable: bool,
    pub decorated: bool,
    pub maximized: bool,
    pub visible: bool,
    pub color: Option<Color>,
    /* private fields */
}

A window.

Fields
title: String

The title of the window.

icon: Option<Image>

The icon of the window.

size: Size

The size of the window.

sizing: WindowSizing

The sizing of the window.

scale: f32

The scale of the window.

Modifying this is not recommended, and will probably not do what you expect, as a rule of thumb, don’t do it.

resizable: bool

Whether the window is resizable.

decorated: bool

Whether the window is decorated.

maximized: bool

Whether the window is maximized.

visible: bool

Whether the window is visible.

color: Option<Color>

The color of the window.

Implementations

impl Window

pub fn new() -> Window

Create a new Window.

pub fn id(&self) -> WindowId

Get the unique identifier of the window.

pub fn title(self, title: impl ToString) -> Window

Set the title of the window.

pub fn icon(self, icon: impl Into<Option<Image>>) -> Window

Set the icon of the window.

pub fn size(self, width: u32, height: u32) -> Window

Set the size of the window.

pub fn sizing(self, sizing: WindowSizing) -> Window

Set the sizing of the window.

pub fn fit_content(self) -> Window

Set the sizing to WindowSizing::Content.

pub fn scale(self, scale: f32) -> Window

Set the scale of the window.

pub fn resizable(self, resizable: bool) -> Window

Set whether the window is resizable.

pub fn decorated(self, decorated: bool) -> Window

Set whether the window is decorated.

pub fn maximized(self, maximized: bool) -> Window

Set whether the window is maximized.

pub fn visible(self, visible: bool) -> Window

Set whether the window is visible.

pub fn color(self, color: impl Into<Option<Color>>) -> Window

Set the color of the window.

pub fn physical_size(&self) -> Size

Get the size of the window in physical pixels.

This is a shorthand for self.size * self.scale.

pub fn width(&self) -> u32

Get the width of the window.

pub fn height(&self) -> u32

Get the height of the window.

pub fn pointers(&self) -> &[Pointer]

Get the pointers in the window.

pub fn pointers_mut(&mut self) -> &mut Vec<Pointer>

Get the pointers in the window mutably.

pub fn is_hovered(&self, view_id: ViewId) -> bool

Get whether a specific view is hovered.

pub fn is_pointer_held( &self, pointer_id: PointerId, button: PointerButton, ) -> bool

Get whether button is held down on pointer with pointer_id.

pub fn get_pointer(&self, pointer_id: PointerId) -> Option<&Pointer>

Get the pointer with pointer_id.

pub fn get_pointer_mut(&mut self, pointer_id: PointerId) -> Option<&mut Pointer>

Get the pointer with pointer_id mutably.

pub fn press_pointer(&mut self, pointer_id: PointerId, button: PointerButton)

Press a button on a pointer.

This is rarely what you want to do, do not use this unless you really know what you are doing.

pub fn release_pointer( &mut self, pointer_id: PointerId, button: PointerButton, ) -> bool

Release a button on a pointer.

This is rarely what you want to do, do not use this unless you really know what you are doing.

pub fn move_pointer(&mut self, pointer_id: PointerId, position: Point) -> Vector

Move a pointer, returning the movement.

pub fn remove_pointer(&mut self, pointer_id: PointerId)

Remove a pointer.

This is rarely what you want to do, do not use this unless you really know what you are doing.

pub fn updates(&mut self) -> Vec<WindowUpdate>

Update the window.

pub fn snapshot(&self) -> WindowSnapshot

Get the WindowSnapshot of the window.