Struct ori::prelude::Window

pub struct Window { /* private fields */ }

A handle to a window.

Implementations

impl Window

pub fn new(raw: Box<dyn RawWindow>, id: WindowId) -> Window

Create a new window with the given raw window implementation.

pub fn downcast_ref<T>(&self) -> Option<&T>
where T: RawWindow,

Try to downcast the window to a specific type.

pub fn id(&self) -> WindowId

Get the WindowId.

pub fn pointers(&self) -> &[Pointer]

Get the pointers.

pub fn pointer(&self, id: PointerId) -> Option<&Pointer>

Get the pointer with id.

pub fn is_hovered(&self, view: ViewId) -> bool

Get the pointer that is currently hovered over the given view.

pub fn downcast_raw<T>(&self) -> Option<&T>
where T: RawWindow,

Try to downcast the window to a specific type.

pub fn title(&self) -> String

Get the title of the window.

pub fn set_title(&mut self, title: &str)

Set the title of the window.

pub fn set_icon(&mut self, icon: Option<&Image>)

Set the icon of the window.

pub fn size(&self) -> Size

Get the size of the window.

pub fn rect(&self) -> Rect

Get the rect of the window.

pub fn physical_size(&self) -> Size

Get the physical size of the window.

pub fn width(&self) -> u32

Get the width of the window.

pub fn height(&self) -> u32

Get the height of the window.

pub fn physical_width(&self) -> u32

Get the physical width of the window.

pub fn physical_height(&self) -> u32

Get the physical height of the window.

pub fn set_size(&mut self, width: u32, height: u32)

Set the size of the window.

pub fn resizable(&self) -> bool

Get whether the window is resizable.

pub fn set_resizable(&mut self, resizable: bool)

Set whether the window is resizable.

pub fn decorated(&self) -> bool

Get whether the window is decorated.

pub fn set_decorated(&mut self, decorated: bool)

Set whether the window is decorated.

pub fn scale_factor(&self) -> f32

Get the scale factor of the window.

pub fn minimized(&self) -> bool

Get whether the window is minimized.

pub fn set_minimized(&mut self, minimized: bool)

Set whether the window is minimized.

pub fn maximized(&self) -> bool

Get whether the window is maximized.

pub fn set_maximized(&mut self, maximized: bool)

Set whether the window is maximized.

pub fn visible(&self) -> bool

Get whether the window is visible.

pub fn set_visible(&mut self, visible: bool)

Set whether the window is visible.

pub fn set_cursor(&mut self, cursor: Cursor)

Set the cursor of the window.

pub fn color(&self) -> Option<Color>

Get the background of the window.

pub fn set_color(&mut self, color: Option<Color>)

Set the background of the window.

pub fn set_soft_input(&mut self, enabled: bool)

Get whether the soft input is enabled.

pub fn drag(&mut self)

Drag the window.

pub fn request_draw(&mut self)

Request a redraw of the window.